requirements.txt
Django==1.10.2
django-filebrowser==3.7.2
django-grappelli==2.8.3
django-registration==2.2
mr-csv-worker==0.0.2
mysqlclient==1.3.9
Pillow==3.4.2
